# HouseParty.Server/openapi/houseparty-server.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "HouseParty.Server | v1",
    "version": "1.0.0"
  },
  "paths": {
    "/api/rooms": {
      "post": {
        "tags": [
          "Rooms"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RoomCreated"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomCreated"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomCreated"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/{roomId}/join": {
      "post": {
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomJoinRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RoomJoinRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RoomJoinRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/RoomJoined"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomJoined"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomJoined"
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/{roomId}/mouse": {
      "post": {
        "tags": [
          "Rooms"
        ],
        "parameters": [
          {
            "name": "roomId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MousePresenceUpdate"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MousePresenceUpdate"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MousePresenceUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/signalr/negotiate": {
      "post": {
        "tags": [
          "SignalR"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/SignalRNegotiation"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalRNegotiation"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignalRNegotiation"
                }
              }
            }
          }
        }
      }
    },
    "/api/engine/turn-based-game/start-game": {
      "post": {
        "tags": [
          "TurnBasedGame"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartGameRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StartGameRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StartGameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StartGameResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartGameResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartGameResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/engine/turn-based-game/start-turn": {
      "post": {
        "tags": [
          "TurnBasedGame"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartTurnRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/StartTurnRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/StartTurnRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/StartTurnResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartTurnResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/StartTurnResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "MousePresenceUpdate": {
        "required": [
          "playerNumber",
          "name",
          "x",
          "y"
        ],
        "type": "object",
        "properties": {
          "playerNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "x": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "y": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "RoomCreated": {
        "required": [
          "id"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          }
        }
      },
      "RoomJoined": {
        "required": [
          "player",
          "players"
        ],
        "type": "object",
        "properties": {
          "player": {
            "$ref": "#/components/schemas/RoomPlayer"
          },
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/RoomPlayer"
            }
          }
        }
      },
      "RoomJoinRequest": {
        "required": [
          "connectionId",
          "name",
          "playerNumber"
        ],
        "type": "object",
        "properties": {
          "connectionId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "playerNumber": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "null",
              "integer",
              "string"
            ],
            "format": "int32"
          }
        }
      },
      "RoomPlayer": {
        "required": [
          "number",
          "name"
        ],
        "type": "object",
        "properties": {
          "number": {
            "pattern": "^-?(?:0|[1-9]\\d*)$",
            "type": [
              "integer",
              "string"
            ],
            "format": "int32"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "SignalRNegotiation": {
        "required": [
          "url",
          "accessToken"
        ],
        "type": "object",
        "properties": {
          "url": {
            "type": "string"
          },
          "accessToken": {
            "type": "string"
          }
        }
      },
      "StartGameRequest": {
        "required": [
          "gameId",
          "playerId"
        ],
        "type": "object",
        "properties": {
          "gameId": {
            "type": "string"
          },
          "playerId": {
            "type": "string"
          }
        }
      },
      "StartGameResponse": {
        "required": [
          "gameStarted"
        ],
        "type": "object",
        "properties": {
          "gameStarted": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "StartTurnRequest": {
        "required": [
          "gameId",
          "playerId"
        ],
        "type": "object",
        "properties": {
          "gameId": {
            "type": "string"
          },
          "playerId": {
            "type": "string"
          }
        }
      },
      "StartTurnResponse": {
        "required": [
          "turnStarted"
        ],
        "type": "object",
        "properties": {
          "turnStarted": {
            "type": "boolean"
          },
          "errorMessage": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "Rooms"
    },
    {
      "name": "SignalR"
    },
    {
      "name": "TurnBasedGame"
    }
  ]
}